various fixes for warnings
git-svn-id: https://svn.osgeo.org/gdal/trunk/gdal@17636 f0d54148-0727-0410-94bb-9a71ac55c965

+++ gcore/gdal_misc.cpp
@@ -1486,7 +1486,7 @@ const char * CPL_STDCALL GDALVersionInfo( const char *pszRequest )
         if( fp != NULL )
         {
             VSIFSeekL( fp, 0, SEEK_END );
-            nLength = VSIFTellL( fp ) + 1;
+            nLength = (int) VSIFTellL( fp ) + 1;
             VSIFSeekL( fp, SEEK_SET, 0 );
 
             pszResultLicence = (char *) VSICalloc(1,nLength);
